January 16 2015 | Window
The window is coming together. I'm able to paint and detect window resizes. It
isn't working perfectly yet, but a window does open.

Adjusted the makefile to use the glfw-cxx library.

Adding support for abstraction of an opengl window on top of the original
Window that is provided by glfw-cxx library. This will make it easier for
drawing in the future. It also looks a lot more like the qt library.