January 18 2015 | File Loader
The polygon files contain an atlas of colours at the top, or bottom.
Technically you could interleave the colours among the vertices in the
polygons, but that would be confusing, hard to read, and dumb.

A polygon can be empty, or have many vertices. A vertex must be defined with at
least two integers representing an x and y coordinate. A third value may be
presented being the index in the colour atlas. The colors are defined in RGB
from 0 - 255. If a vertex is defined before a polygon is defined, it will imply
a first polygon. If no colours are defined, then the polygon will be drawn in
white.

The fileloader appears to be working, though it doesn't save the data to the
world yet. It reads the data in from the poly file and places information into
the corresponding data structures. Either polygons or colors.

Testing will occur later.