January 19 2015 | World and drawing library
I'm adding a quick opengl drawing library to the project to make drawing shapes
easier. It draws lines and circles.

I've started implementing the world. I have to fix the opengl viewport. Right
now it does a lovely job seg-faulting because it is trying to close a
non-existent window. This is known behaviour and will be fixed. I think that
I'm going to be multi-threading it to handle changes in the system, such as
when the world is changed.