January 26 2015 | Circle Intersect and Redrawing
I've updated the glviewport class a little so it handles the escape key and x
button correctly. I've also allowed it to redraw and resize the window. It
currently redraws on every cycle of the main for-loop, which will be optimized.

I've also started adding a triangle which will be able to be drawn directly