January 27 2015 | Triangulation
Added some colors to the polygons to show that the triangulations are good.

Added z, x, and c keys to modify properties of the viewport. Z sets the
generated polygon to wireframe view. X toggles the lines, and C toggles the
vertices.

I've moved all of the stuff from the Delaunay files to the world.

I still haven't connected the last triangle, but the rest appear to be building
correctly. The triangulations appear to be fairly smooth.