January 30 2015 | New makefile
This is probably my most powerful makefile yet.

It compiles c, C, cc, and cpp files.
Headers and source files are separated.
Source files are in src/
Header files are in inc/

It can compile everything. (all target)
It can compile with optimizations (for release target)
It can compile with debug -DDEBUG defined and with function names for gdb
(debug target)

It can compile the unit tests in the test/ directory. (test target)