February 4 2015 | GLFW window
I've started writing the header and implementation for the window. I'm trying
to hide the C as much as I can and make it behave similarly to the system in
Qt.

I removed the linear algebra code from the main file. I've decided to go with
glm, so I won't have to worry about my own library and some of the more
advanced functionality.