February 6 2015 | Events and world
I've started implementing the world on top of the window. The world captures
key presses, released, and repeats. It also captures mouse button presses and
released.

I've also started working on making the world be able to draw. I've added the
ability to draw the grid, but there is no camera control so it looks like a
straight line across the window.