camera matrices
The camera matrices are finally making sense and giving a good output.

The camera still doesn't have the ability to move, but it will soon.

I'm going to convert all the vectors in the camera to homogeneous ones so I
don't have to call the conversion function every camera operation.

I may also cache the VP matrix so it doesn't have to be recalculated each
frame.