Starting on marching triangles polygonizer
I've started implementing the marching triangles polygonizer

I've started with a basic seed triangle using the tangent space generated by
the normal of the start point. This method is non-optimal. I would like to find
the direction of the most curvature and move in that direction. Then use the
midpoint of the line as a way to find the next two vertices, or four edges.

From there, the algorithm will take over.