Switching to Implicit Library
Instead of storing the implicit system in the polygonizer, we are going to link
to the implicit system project.

Please build the implicit system. In the release file, there will be a file for
all of the headers. Copy these into 'inc/implicit/' directory. Copy the library
libImplicit.a to 'lib/'.