Starting Polygonizer Implementation
I'm not sure that I completely like where this is going, but I've started
digging into the meaty bits of the polygonization. I want to do some renaming
though because the names are too long and bulky. Maybe some separation of the
components into separate files for the purpose of clarity. In the future, we
should be able to shrink the entire thing down again, but for now I would like
to rename and clean it up a bit.

I'm basing the polygonization algorithm off of the method described in
"Curvature dependent Polygonization of Implicit Surfaces" by Bruno Rogrigues de
Araujo and Joaquim Armando Pires Jorge.