Split Polygonizer
I've split up the polygonizer into many classes. I've also added some helper
functions in the utilities folders.

Front: The front is the outermost edge of the polygonization. This is where
growth occurs.

Edge: An edge is made up of two vertices, not too exciting.

PolyTris: This is where the polygonizer actually lives. I'll be implementing
the actual algorithm here.

GeoCore: Contains helpful functions for use with geometric things. This allows
the calculation of an angle between vectors, finding the tangent space. It will
contain any helper function for working with geometry.

MathCore: This contains helper functions for use with math types. For now it
contains the clamp function, which bounds a value from above and below.