Added a pop vertex to the front
I've implemented a popVertex to the front class. This removes a vertex from the
front and updates the neighbor information for a given front vertex.

I may add a new functionality that just kills the meaning of neighbors since
the vertex will be removed when a new vertex builds on top of it and removes it
from the front.

We may need to adjust the implementation of the vertex structure though.