Updates to the PolyFront
The front has been updated to keeping the data elements in order.
By keeping the elements in order, we can simplify the implementation of the
front vertex. I have removed the front vertex, and have replaced it with three
vector lists containing the vertex index, the opening angle, and the radius of
curvature for a vertex at some position in the front. Keeping the left neighbor
in the next index smaller than our current index, and the right neighbor in the
next index greater than our current index, we keep the list in order and no
longer require the neighbor index values as they will always be one away from
the current index.