Index Fixing in PolyFront
I've fixed a lot of the indexing issues. The main issue was the subtraction off
the size in the modulus operation. It should not have been subtracted.

The left index operation was having issues because we are using unsigned int
indexing.  0 - 1 % list_size does not return the last element of the list.
I've added a quick if-statement to handle this case.

The PolyContainer has been updated. I've made the members private.
We need to store a map from a glm::vec3 to a vertex index. For this reason, it
is necessary that all insertions to the vertex list be through a method call.
For consistency purposes, I've made all members private, but fully accessible
through the method calls. I have also updated the rest of the program to handle
this change.

Finally, changes have been made to better reflect the psuedo-code algorithm
provided in the paper we are implementing.