Front Kd Tree Working
The Front kd tree can test for potential self-intersections.
Under normal circumstances, there should be 1, 2, or 3 vertices in the search
results. If there is 1, it is the vertex we are looking at. If there are 2 and
the other is our neighbor, ignore it. If there are three, both must be our
neighbors to be under normal circumstances. If there are more than 3, or the
other vertex is not our neighbor, connect and split fronts.